#11, separo objetos valor de la entidad producto
