+++ internal/producto.go
@@ -5,30 +5,6 @@ import (
 	"fmt"
 )
 
-type Precio struct {
-	Valor float64
-}
-
-func NuevoPrecio(valor float64) (Precio, error) {
-	if valor < 0 {
-		return Precio{}, fmt.Errorf("El precio no puede ser negativo")
-	}
-	return Precio{Valor: valor}, nil
-}
-
-
-type Alergia struct {
-	Nombre string
-}
-
-func NuevaAlergia(nombre string) (Alergia, error) {
-	if nombre == "" {
-		return Alergia{}, fmt.Errorf("La alergia debe tener nombre")
-	}
-	return Alergia{Nombre: nombre}, nil
-}
-
-
 type Producto struct {
 	Nombre   string
 	Precio   Precio
